v1.10.190 — Build-Fix: MagickGeometry int statt uint
v1.10.189 hatte versehentlich (uint)size-Casts drin. Magick.NET 13.10
(installiert) nutzt int für MagickGeometry — der 14.x-Aufruf failed mit
CS1503. Cast raus, sonst identisch zu .189.

+++ src/NimShare.Api/Services/ThumbnailService.cs
@@ -163,7 +163,7 @@ public class ThumbnailService : IThumbnailService
 
                 img.AutoOrient();
                 img.Strip();
-                img.Resize(new MagickGeometry((uint)size, (uint)size)
+                img.Resize(new MagickGeometry(size, size)
                 {
                     Greater = true,
                 });
